Removed single-touch check from endTouch
Making a check for a single-fingered touch in endTouch was causing sling-shot behavior if the timing of lifting the fingers was off. As our use-case would never involve transitioning between a two-finger touch directly into a one-finger touch (without lifting the fingers), eliminating this check was a quick way to solve the sling-shot problem.